Добавлен компонент src/components/comments.js
