+++ public/css/main.css
@@ -531,7 +531,7 @@ button {
   top: 0;
   right: 0;
   left: 10%;
-  z-index: 1;
+  z-index: 10;
   background-color: #212126;
   box-shadow: -3px 2px 13px 0 #141628;
   overflow-x: hidden;
