+++ src/components/film-details.js
@@ -1,6 +1,5 @@
 import AbstractSmartComponent from "./abstract-smart-component";
-import {formatTimeHour, formatTimeMinute, formatDate, formatDateComment} from "../utils/common";
-
+import {formatTimeHour, formatTimeMinute, formatDate} from "../utils/common";
 
 const createGenreMarkup = (genres) => {
   return genres.map((genre) => {
@@ -11,29 +10,7 @@ const createGenreMarkup = (genres) => {
   .join(`\n`);
 };
 
-const createCommmentsMarkup = (comments) => {
-  return comments.map((comment) => {
-    const commentTimeAgo = formatDateComment(comment.data);
-    return (
-      `<li class="film-details__comment">
-        <span class="film-details__comment-emoji">
-          <img src="${comment.emotion}" width="55" height="55" alt="emoji-smile">
-        </span>
-        <div>
-          <p class="film-details__comment-text">${comment.сomment}</p>
-          <p class="film-details__comment-info">
-            <span class="film-details__comment-author">${comment.avtor}</span>
-            <span class="film-details__comment-day">${commentTimeAgo}</span>
-            <button class="film-details__comment-delete">Delete</button>
-          </p>
-        </div>
-      </li>`
-    );
-  });
-};
-
-
-const createFilmDetailsTemplate = (film, options = {}) => {
+const createFilmDetailsTemplate = (film) => {
   const {
     film_info: {
       title,
@@ -57,10 +34,7 @@ const createFilmDetailsTemplate = (film, options = {}) => {
       already_watched: alreadyWatched,
       favorite,
     },
-    comments,
   } = film;
-
-  const {newElementImgEmojiSrc, newElementImgEmojiAlt, resetTextariaEmojValue} = options;
 
   const writersDetails = writers.join(`, `);
   const actorsDetails = actors.join(`, `);
@@ -72,8 +46,6 @@ const createFilmDetailsTemplate = (film, options = {}) => {
 
   const genreDeyails = genre.length > 1 ? `Genres` : `Genre`;
   const genreMarkup = createGenreMarkup(genre);
-  const commentsNumber = comments.length;
-  const CommmentsMarkup = createCommmentsMarkup(comments);
 
   return (
     `<section class="film-details">
@@ -152,54 +124,11 @@ const createFilmDetailsTemplate = (film, options = {}) => {
           </section>
         </div>
 
-        <div class="form-details__bottom-container">
-          <section class="film-details__comments-wrap">
-            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${commentsNumber}</span></h3>
 
-            <ul class="film-details__comments-list">
-             ${CommmentsMarkup}
-            </ul>
-
-            <div class="film-details__new-comment">
-              <div for="add-emoji" class="film-details__add-emoji-label">
-                <img src="${newElementImgEmojiSrc}" width="55" height="55" alt="${newElementImgEmojiAlt}">
-              </div>
-
-              <label class="film-details__comment-label">
-                <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment">${resetTextariaEmojValue}</textarea>
-              </label>
-
-              <div class="film-details__emoji-list">
-                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile" checked>
-                <label class="film-details__emoji-label" for="emoji-smile">
-                  <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
-                </label>
-
-                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-                <label class="film-details__emoji-label" for="emoji-sleeping">
-                  <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
-                </label>
-
-                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-                <label class="film-details__emoji-label" for="emoji-puke">
-                  <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
-                </label>
-
-                <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-                <label class="film-details__emoji-label" for="emoji-angry">
-                  <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
-                </label>
-              </div>
-            </div>
-          </section>
-        </div>
       </form>
     </section>`
   );
 };
-// <div for="add-emoji" class="film-details__add-emoji-label">
-//    <img src="images/emoji/smile.png" width="55" height="55" alt="emoji-smile">
-// </div>
 
 export default class FilmDetails extends AbstractSmartComponent {
   constructor(film) {
@@ -207,21 +136,10 @@ export default class FilmDetails extends AbstractSmartComponent {
 
     this._film = film;
     this._submitHandler = null;
-
-    this._newElementImgEmojiSrc = ``;
-    this._newElementImgEmojiAlt = ``;
-    this._resetTextariaEmojValue = ``;
-
-    this._subscribeOnEvents();
   }
 
   getTemplate() {
-    return createFilmDetailsTemplate(this._film, {
-      newElementImgEmojiSrc: this._newElementImgEmojiSrc,
-      newElementImgEmojiAlt: this._newElementImgEmojiAlt,
-      resetTextariaEmojValue: this._resetTextariaEmojValue,
-    });
-
+    return createFilmDetailsTemplate(this._film);
   }
 
   recoveryListeners() {
@@ -234,8 +152,6 @@ export default class FilmDetails extends AbstractSmartComponent {
   }
 
   reset() {
-    this._clear();
-
     this.rerender();
   }
 
@@ -244,12 +160,6 @@ export default class FilmDetails extends AbstractSmartComponent {
       .addEventListener(`click`, handler);
 
     this._submitHandler = handler;
-  }
-
-  _clear() {
-    this._newElementImgEmojiSrc = ``;
-    this._newElementImgEmojiAlt = ``;
-    this._resetTextariaEmojValue = ``;
   }
 
   _subscribeOnEvents() {
@@ -273,21 +183,6 @@ export default class FilmDetails extends AbstractSmartComponent {
       this._film[`user_details`][`favorite`] = !this._film[`user_details`][`favorite`];
       // this.rerender();
     });
-
-    element.querySelector(`.film-details__emoji-list`).addEventListener(`click`, (evt) => {
-      const target = evt.target;
-      const elementImgEmoji = target.closest(`img`);
-      if (target && elementImgEmoji) {
-        this._newElementImgEmojiSrc = elementImgEmoji.src;
-        this._newElementImgEmojiAlt = elementImgEmoji.alt;
-      }
-      this.rerender();
-    });
-
-    element.querySelector(`.film-details__comment-input`).addEventListener(`change`, (evt) => {
-      this._resetTextariaEmojValue = evt.target.value;
-    });
-
 
   }
